Add test button functionality
