+++ src/components/NavigationMenu/NavigationMenu.scss
