+++ src/styles/globals.scss
@@ -1,0 +1,37 @@
+*,
+::before,
+::after {
+  box-sizing: border-box;
+}
+
+body {
+  background-color: #e5e5e5;
+  margin: 0 auto;
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+ul,
+li {
+  display: flex;
+  align-items: stretch;
+  gap: 10px;
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+h1,
+h2 {
+  margin: 0;
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+button {
+  background: none;
+  border: 0;
+  cursor: pointer;
+}
