+++ src/styles/index.scss
@@ -1,0 +1,1 @@
+@use "./globals.scss";
